task/requirements.txt
asgiref==3.9.1
Django==5.2.6
sqlparse==0.5.3
djangorestframework==3.16.1
dkimpy==1.1.8
dnspython==2.8.0
drf-yasg==1.21.10
inflection==0.5.1
packaging==25.0
pytz==2025.2
PyYAML==6.0.2
uritemplate==4.2.0
djangorestframework-simplejwt==5.3.1  
cryptography==42.0.5    
celery==5.3.6                
redis==5.0.1
drf-yasg==1.21.10                               
